Input checks for cashier.cpp
Added check for char in addBookMenu().

Added option to exit from entering an identifier to cart. Previously you
would get stuck in the menu with no way to get back to the main menu
other than closing the program. This is currently set to 0 but it can be
changed to any value you guys want. It's all I could think of since the
program already tracks -1 for no ID found. I don't believe 0 is ever
used as an identifier since the default constructor sets it to 1. Change
it if I'm wrong please!

Changed some output text

+++ CIS22B_FinalProject/CIS22B_FinalProject/Cashier.cpp
@@ -20,62 +20,74 @@ void Cashier::addBookMenu()
 	bool correctInput = false;
 	while (correctInput == false)
 	{
-		cout << "Enter the identifier of the book you wish to add to cart: ";
-		cin >> identifier;
-		Book* tempBook = database->searchIdentifier(identifier);
-		if (tempBook->getIdentifier() == -1)
+		cout << "Enter the identifier of the book you wish to add to cart or 0 to exit: ";
+		if (cin >> identifier)
 		{
-			cout << "A book with the identifier " << identifier << " was not found.\n";
-			cout << "Press return to continue.";
-			cin.ignore();
-			cin.get();
-		}
-		else
-		{
-			cout << "Does this look correct?\n\n" << *tempBook << endl;
-			cout << "Response (Y/N)> ";
-			string response;
-			cin >> response;
-			if (response == "y" || response == "Y")
+			Book* tempBook = database->searchIdentifier(identifier);
+			if (identifier == 0)
 			{
-				bool bookAlreadyInCart = false;
-				for (int i = 0; i < cartSize; i++)
-				{
-					if (cart[i] == identifier)
-					{
-						bookAlreadyInCart = true;
-					}
-				}
-				if (bookAlreadyInCart)
-				{
-					cout << "That book is already in the cart.\n";
-					cout << "Press return to continue.";
-					cin.ignore();
-					cin.get();
-				}
-				else
-				{
-					correctInput = true;
-					cart[cartSize] = identifier;
-					cartSize++;
-					cout << "The book " << tempBook->getTitle() << " was added to the cart.\n";
-					cout << "Press return to continue.";
-					cin.ignore();
-					cin.get();
-				}
+				correctInput = true;
 			}
-			else if (response == "n" || response == "N")
+			else if (tempBook->getIdentifier() == -1)
 			{
-				cout << "\nPlease try again.\n\n";
-				cin.ignore();
-			}
-			else
-			{
-				cout << "Invalid response...\n";
+				cout << "A book with the identifier " << identifier << " was not found.\n";
 				cout << "Press return to continue.";
 				cin.ignore();
 				cin.get();
 			}
+			else
+			{
+				cout << "Does this look correct?\n\n" << *tempBook << endl;
+				cout << "Response (Y/N)> ";
+				string response;
+				cin >> response;
+				if (response == "y" || response == "Y")
+				{
+					bool bookAlreadyInCart = false;
+					for (int i = 0; i < cartSize; i++)
+					{
+						if (cart[i] == identifier)
+						{
+							bookAlreadyInCart = true;
+						}
+					}
+					if (bookAlreadyInCart)
+					{
+						cout << "That book is already in the cart.\n";
+						cout << "Press return to continue.";
+						cin.ignore();
+						cin.get();
+					}
+					else
+					{
+						correctInput = true;
+						cart[cartSize] = identifier;
+						cartSize++;
+						cout << "The book " << tempBook->getTitle() << " was added to the cart.\n";
+						cout << "Press return to continue.";
+						cin.ignore();
+						cin.get();
+					}
+				}
+				else if (response == "n" || response == "N")
+				{
+					cout << "\nPlease try again.\n\n";
+					cin.ignore();
+				}
+				else
+				{
+					cout << "Invalid response. Please try again\n";
+					cout << "Press return to continue.";
+					cin.ignore();
+					cin.get();
+				}
+			}
+		}
+		else
+		{
+			cout << "Invalid response. Please try again.\n\n";
+			cin.clear();
+			cin.ignore(1000, '\n');
 		}
 	}
 }
@@ -191,46 +203,56 @@ void Cashier::checkout()
 {
 	system("CLS");
 	double subtotal = 0;
-	for (int i = 0; i < cartSize; i++)
-	{
-		subtotal += database->getPrice(cart[i]);
-	}
-	cout << "Serendipity Book Sellers" << endl << endl;
-	cout << "Date: " << endl << endl;
-	cout << "ISBN\t\tTitle\t\t\t\tPrice" << endl;
-	cout << "____________________________________________________________" << endl << endl;
-	for (int i = 0; i < cartSize; i++)
-	{
-		Book* book = database->searchIdentifier(cart[i]);
-		cout << book->getIsbn() << "\t\t" << book->getTitle() << "\t\t\t\t" << book->getRetailPrice() << endl;
-	}
-	cout << "\t\t\t\t\t____________________" << endl << endl;
-	cout << "\t\tSubtotal:\t\t\t" << subtotal << endl << "\t\tTax: " << "Sales Tax @ 6.25%:\t\t" << (subtotal * 0.0625) << endl << "\t\tTotal:\t\t\t\t" << subtotal + subtotal*0.0625 << endl << endl;
-	cout << "Does this look correct? (Y/N)> ";
-	string response;
-	cin >> response;
-	if (response == "y" || response == "Y")
+	if (cartSize > 0)
 	{
 		for (int i = 0; i < cartSize; i++)
 		{
-			database->removeBook(cart[i]);
+			subtotal += database->getPrice(cart[i]);
 		}
-		cartSize = 0;
-		cout << "Transaction successful.\n";
-		cout << "Press return to continue.";
-		cin.ignore();
-		cin.get();
-	}
-	else if (response == "n" || response == "N")
-	{
-		cout << "Exiting...\n";
-		cout << "Press return to continue.";
-		cin.ignore();
-		cin.get();
+		cout << "Serendipity Book Sellers" << endl << endl;
+		cout << "Date: " << endl << endl;
+		cout << "ISBN\t\tTitle\t\t\t\tPrice" << endl;
+		cout << "____________________________________________________________" << endl << endl;
+		for (int i = 0; i < cartSize; i++)
+		{
+			Book* book = database->searchIdentifier(cart[i]);
+			cout << book->getIsbn() << "\t\t" << book->getTitle() << "\t\t\t\t" << book->getRetailPrice() << endl;
+		}
+		cout << "\t\t\t\t\t____________________" << endl << endl;
+		cout << "\t\tSubtotal:\t\t\t" << subtotal << endl << "\t\tTax: " << "Sales Tax @ 6.25%:\t\t" << (subtotal * 0.0625) << endl << "\t\tTotal:\t\t\t\t" << subtotal + subtotal*0.0625 << endl << endl;
+		cout << "Does this look correct? (Y/N)> ";
+		string response;
+		cin >> response;
+		if (response == "y" || response == "Y")
+		{
+			for (int i = 0; i < cartSize; i++)
+			{
+				database->removeBook(cart[i]);
+			}
+			cartSize = 0;
+			cout << "Transaction successful.\n";
+			cout << "Press return to continue.";
+			cin.ignore();
+			cin.get();
+		}
+		else if (response == "n" || response == "N")
+		{
+			cout << "Exiting...\n";
+			cout << "Press return to continue.";
+			cin.ignore();
+			cin.get();
+		}
+		else
+		{
+			cout << "Invalid response. Please try again.\n";
+			cout << "Press return to continue.";
+			cin.ignore();
+			cin.get();
+		}
 	}
 	else
 	{
-		cout << "Invalid response...\n";
+		cout << "\n\n\t    The cart is empty!\n\n";
 		cout << "Press return to continue.";
 		cin.ignore();
 		cin.get();
